test(event): Added tests for each correctly set up rule test(date): Check for invalid constructor initial values refactor: Removed unnecesary `const` modifiers refactor: Added const modifiers to parameters which call `Date` objects refactor(date): Operators now get the address instead copy refactor(event): `print()` now use `Date::toString()` refactor(regex): moved all regexes to `event.h` refactor(rule/fixed-range): Stores `Date` objects instead of pointers fix(event): Now adds tags / descriptions correctly fix(event/rules): memory ownership for rules made in constructor fix(rule/day-of-week): Correctly assign value when converting from char build: Treat Warnings as Errors build: Refactored Makefile to speed up compilation

+++ strings.cpp
@@ -12,7 +12,7 @@ int_least16_t str_find_tag_clousure(std::string str, char opening, char closure)
 	//Searches for [...] at the start of description, returns -1 if NOT found, otherwise the index of the ']' character
 	//also returns -1 when the initial character isn't tag opening
 	if (str[0] != opening) return -1;
-	int_least16_t tagClosureIndex = str.find_first_of(closure);
+	size_t tagClosureIndex = str.find_first_of(closure);
 	if (tagClosureIndex == str.npos) return -1;
 	return tagClosureIndex;
 }
